#4 refactor : 폴더 구조 web, application, domain, infrastructure로 구분
