+++ geulkkoli/src/main/java/com/geulkkoli/infrastructure/EmptyDataException.java
@@ -1,4 +1,4 @@
-package com.geulkkoli.respository;
+package com.geulkkoli.infrastructure;
 
 public class EmptyDataException extends RuntimeException{
     public EmptyDataException(String message) {
